requirements.txt
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.5
chromadb>=0.5.15
pypdf>=4.3.1
sentence-transformers>=3.1.1
streamlit>=1.38.0
pydantic>=2.9.2
tiktoken>=0.7.0
python-dotenv>=1.0.0
